Refactoring: renamed ClientProviders, created user context, tweaks to login and navbar
