+++ apps/frontend/src/app/layout.tsx
@@ -1,6 +1,7 @@
 import "../styles/globals.css";
 import { Inter } from "next/font/google";
-import ClientProviders from "@/components/ClientProvider/ClientProviders";
+import ClientProviders from "@/components/Provider/ClientProviders";
+import {UserProvider} from "@/context/UserContext";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -17,7 +18,9 @@ export default function RootLayout({
   return (
     <html lang="en" suppressHydrationWarning>
       <body className={inter.className}>
+        <UserProvider>
         <ClientProviders>{children}</ClientProviders>
+        </UserProvider>
       </body>
     </html>
   );
