+++ apps/frontend/src/services/auth.ts
@@ -6,6 +6,7 @@ export interface LoginResponse {
   user: {
     id: number;
     email: string;
+    name: string;
   };
 }
